requirements.txt
Django~=5.1
djangorestframework~=3.15.2
psycopg2-binary
python-decouple>=3.6
